Add windows packaging
Changed API of the C++ bridge: new take a reference to a std::string in
which the deserialized qbjs content is copied. The original idea of
returning a rust::String didn't work because of some missing destructor
symbols on Windows, so we use this alternative. In a way it is better as
the C++ only handle C++ types now.